# swagger.yaml
openapi: 3.0.1
info:
  title: Movie Critics Api
  description: "Movie Critics Api - provides access to movie data and allows you to create lists with movies, comment on movies, and set ratings."
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: /
tags:
- name: Auth
  description: "Authentication endpoints"
- name: Comments
  description: "Operations for creating, editing and deleting comments"
- name: Films
  description: "Operations for getting: film, filters, comments; rating films"
- name: Lists
  description: "Operations for creating, editing and deleting lists"
  
paths:
  /api/auth/register:
    post:
      tags:
      - Auth
      summary: "Register new user"
      operationId: "registerUser"
      requestBody:
        description: "User data for registration"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDTO"
        required: true
      responses:
        "201":
          description: "User created"
        "400":
          description: "Invalid input"
      
  /api/auth/login:
    post:
      tags:
      - Auth
      summary: "Login user"
      description: " Logs in the user and returns a session."
      operationId: "loginUser"
      requestBody:
        description: "User login data"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDTO"
        required: true
      responses:
        "200":
          description: "User logged in"
          headers:
            Set-Cookie:
              description: "Session tokens are set as cookies. Example: `jwt=abcd1234; 
                \ HttpOnly; Path=/; Max-Age=3600` and `refresh_jwt=abcd5678; HttpOnly;\
                \ Path=/; Max-Age=86400`"
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEntity"
        "401":
          description: 'Unauthorized'
          
  /api/auth/logout:
    get:
      tags:
      - Auth
      summary: "Logout user"
      operationId: "logoutUser"
      responses:
        "200":
          description: "User logged out"
        "401":
          description: "Unauthorized"
      security:
      - AccessToken: []
      
  /api/auth/:
    get:
      tags:
      - Auth
      summary: "Get authenticated user data"
      operationId: "getUserData"
      responses:
        "200":
          description: "User authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEntity"
        "401":
          description: "Unauthorized"
      security:
      - AccessToken: []
      
  /api/auth/refresh:
    get:
      tags:
      - Auth
      summary: "Refresh access token"
      operationId: "refreshToken"
      responses:
        "200":
          description: "Access token updated"
          headers:
            Set-Cookie:
              description: "Access token are set as cookie. Example: `refresh_jwt=abcd5678;\
                \ HttpOnly; Path=/; Max-Age=86400`"
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEntity"
        "401":
          description: "Unauthorized"
      security:
      - RefreshToken: []
      
  /api/films:
    get:
      tags:
      - Films
      summary: "Get films data"
      description: "Fetches a paginated list of films."
      operationId: "getFilms"
      parameters:
      - name: page
        in: query
        description: "Page number for pagination"
        schema:
          type: integer
          default: 1
      - name: limit
        in: query
        description: "Number of films per page"
        schema:
          type: integer
          default: 10
      responses:
        "200":
          description: "List of films"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmsEntity"
                
  /api/films/filters:
    get:
      tags:
      - Films
      summary: "Get filters data"
      operationId: "getFilters"
      responses:
        "200":
          description: "Filters data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FiltersEntity"
                
  /api/films/{id}:
    get:
      tags:
      - Films
      summary: "Get film data by ID"
      description: "Fetches detailed information about a specific film by its ID."
      operationId: "getFilmById"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          example: 1
      responses:
        200:
          description: "Film data with extras"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmWithExtrasEntity"
        404:
          description: "Film not found"
          
  /api/films/{id}/rate:
    post:
      tags:
      - Films
      summary: "Rate a film"
      operationId: "rateFilm"
      security:
      - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateFilmBodyDTO"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: "integer"
      responses:
        200:
          description: "Film rated successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingEntity"
        401:
          description: "Unauthorized"
          
  /api/films/{id}/comments:
    get:
      tags:
      - Films
      summary: "Get comments by film ID"
      description: "Retrieves comments for a specific film by its ID."
      operationId: "getCommentsByFilmId"
      parameters:
      - name: id
        in: path
        required: true
        description: "ID of the film to retrieve comments for"
        schema:
          type: integer
      responses:
        200:
          description: "List of comments for the film"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentEntity"
        404:
          description: "Film not found"
  
  /api/comments:
    post:
      tags:
      - Comments
      summary: "Create comment"
      operationId: "createComment"
      security:
      - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentDTO"
      responses:
        200:
          description: "Comment created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentEntity"
        401:
          description: "Unauthorized"
          
  /api/comments/{id}:
    delete:
      tags:
      - Comments
      summary: "Delete comment by ID"
      operationId: "deleteComment"
      security:
      - AccessToken: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        200:
          description: "Comment deleted successfully"
        401:
          description: "Unauthorized"
        404:
          description: "Comment not found"
    patch:
      tags:
      - Comments
      summary: "Edit comment by ID"
      operationId: "editComment"
      security:
      - AccessToken: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditCommentDTO"
      responses:
        200:
          description: "Comment edited successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentEntity"
        401:
          description: "Unauthorized"
        404:
          description: "Comment not found"        
          
  /api/lists:
    get:
      tags:
      - Lists
      summary: "Get all public lists"
      operationId: "getLists"
      responses:
        200:
          description: "List of public lists"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShortInfoListWithAuthorAndFilms"
    post:
      tags:
      - Lists
      summary: "Create list"
      operationId: "createList"
      security:
      - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListDTO"
      responses:
        200:
          description: "List created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEntity"
        401:
          description: "Unauthorized"
          
  /api/lists/my:
    get:
      tags:
      - Lists
      summary: "Get authentificated user lists"
      operationId: "getMyLists"
      security:
      - AccessToken: []
      responses:
        200:
          description: "User's lists"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShortInfoListWithAuthorAndFilms"
        401:
          description: "Unauthorized"
          
  /api/lists/{id}:
    get:
      tags:
      - Lists
      summary: "Get list by ID"
      description: "Returns public list or if user authentificated return his private list"
      operationId: "getListById"
      security:
      - AccessToken: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        200:
          description: "List data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfoListWithAuthorAndFilms"
        404:
          description: "List not found"
    delete:
      tags:
      - Lists
      summary: "Delete list by ID"
      operationId: "removeList"
      security:
      - AccessToken: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        200:
          description: "List deleted successfully"
        401:
          description: "Unauthorized"
        404:
          description: "List not found"
    patch:
      tags:
      - Lists
      summary: "Change list visibility"
      operationId: "changeListVisibility"
      security:
      - AccessToken: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeListVisibilityDTO"
      responses:
        200:
          description: "List visibility updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEntity"
        401:
          description: "Unauthorized"
        404:
          description: "List not found"
          
  /api/lists/{id}/films/{filmId}:
    post:
      tags:
      - Lists
      summary: "Add film to list"
      operationId: "addFilmToList"
      security:
      - AccessToken: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: filmId
        in: path
        required: true
        schema:
          type: integer
      responses:
        200:
          description: "Film added to list"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmInListEntity"
        401:
          description: "Unauthorized"
        404:
          description: "List or film not found"
    delete:
      tags:
      - Lists
      summary: "Remove film from list"
      operationId: "removeFilmFromList"
      security:
      - AccessToken: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: filmId
        in: path
        required: true
        schema:
          type: integer
      responses:
        200:
          description: "Film removed from list"
        401:
          description: "Unauthorized"
        404:
          description: "List or film not found"
                
components:
  schemas:
    RegisterDTO:
      type: object
      properties:
        username:
          type: string
          example: user123
        password:
          type: string
          example: password123
        email:
          type: string
          example: user@example.com
    
    LoginDTO:
      type: object
      properties:
        username:
          type: string
          example: user123
        password:
          type: string
          example: password123
    
    UserEntity:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: user123
        email:
          type: string
          example: pXwQ8@example.com
        role:
          type: string
          example: CLIENT
          enum:
          - CLIENT
          - ADMIN
          - ROOT
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T12:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-02T12:00:00Z
          
    FilmEntity:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier of the film
          example: 1282979
        nameRu:
          type: string
          example: LEGO Ниндзяго
          nullable: true
        nameOriginal:
          type: string
          example: null
          nullable: true
        slogan:
          type: string
          example: It's all fun and games until you reach Level 13.
        description:
          type: string
          example: Ниндзя сражаются со злодеями в Ниндзяго-Сити и за его пределами...
        shortDescription:
          type: string
          example: Заскучавшие ниндзя случайно выпускают из заточения опасную злодейку.
        filmLength:
          type: integer
          description: Length of the film in minutes or in series
          example: 11
        year:
          type: integer
          example: 2019
        posterUrl:
          type: string
          example: https://placehold.co/600x400
        posterUrlPreview:
          type: string
          example: https://placehold.co/600x400
        createdAt:
          type: string
          format: date-time
          example: 2023-11-07T12:00:00Z
        type:
          type: string
          example: TV_SERIES
          enum:
          - VIDEO
          - FILM
          - MINI_SERIES
          - TV_SERIES
          - TV_SHOW
        countries:
          type: array
          example:
          - name: США
          - name: Канада
          - name: Дания
          items:
            type: object
            properties:
              name:
                type: string
        genres:
          type: array
          example:
          - name: приключения
          - name: боевик
          - name: мультфильм
          items:
            type: object
            properties:
              name:
                type: string
        avgRating:
          type: number
          format: float
          example: 6.0
          
    FilmsEntity:
      type: object
      properties:
        totalItems:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
        page:
          type: integer
          example: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/FilmEntity'
            
    FilmWithExtrasEntity:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier of the film
          example: 1282979
        nameRu:
          type: string
          example: LEGO Ниндзяго
          nullable: true
        nameOriginal:
          type: string
          example: null
          nullable: true
        slogan:
          type: string
          example: It's all fun and games until you reach Level 13.
        description:
          type: string
          example: Ниндзя сражаются со злодеями в Ниндзяго-Сити и за его пределами...
        shortDescription:
          type: string
          example: Заскучавшие ниндзя случайно выпускают из заточения опасную злодейку.
        filmLength:
          type: integer
          description: Length of the film in minutes or in series
          example: 11
        year:
          type: integer
          example: 2019
        posterUrl:
          type: string
          example: https://placehold.co/600x400
        posterUrlPreview:
          type: string
          example: https://placehold.co/600x400
        createdAt:
          type: string
          format: date-time
          example: 2023-11-07T12:00:00Z
        type:
          type: string
          example: TV_SERIES
          enum:
          - VIDEO
          - FILM
          - MINI_SERIES
          - TV_SERIES
          - TV_SHOW
        countries:
          type: array
          example:
          - name: США
          - name: Канада
          - name: Дания
          items:
            type: object
            properties:
              name:
                type: string
        genres:
          type: array
          example:
          - name: приключения
          - name: боевик
          - name: мультфильм
          items:
            type: object
            properties:
              name:
                type: string
        ratings:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 1
              userRating:
                type: integer
                example: 6
              userId:
                type: integer
                example: 1
              filmId:
                type: integer
                example: 1
        avgRating:
          type: number
          format: float
          example: 6.0
        boxOffice:
          type: string
          nullable: true
          example: '$ 100,000,000'
        awards:
          type: string
          nullable: true
          example: '3 wins'
            
    FiltersEntity:
      type: object
      properties:
        countries:
          type: array
          example:
          - США
          - Канада
          - Дания
          items:
            type: string
        genres:
          type: array
          example:
          - приключения
          - боевик
          - мультфильм
          - семейный
          - детский
          items:
            type: string
        types:
          type: array
          example:
          - VIDEO
          - FILM
          - MINI_SERIES
          - TV_SERIES
          - TV_SHOW
          items:
            type: string
            
    RateFilmBodyDTO:
      type: object
      properties:
        rating:
          type: integer
          description: "Rating value for the film, typically between 1 and 5"
          example: 4
          
    RatingEntity:
      type: object
      properties:
        id:
          type: integer
          example: 1
        filmId:
          type: integer
          example: 1
        userId:
          type: integer
          example: 1
        userRating:
          type: integer
          example: 5
          
    CreateCommentDTO:
      type: object
      properties:
        content:
          type: string
          description: The content of the comment
          example: This is a new comment.
          
    EditCommentDTO:
      type: object
      properties:
        content:
          type: string
          description: The updated content of the comment
          example: Updated comment content.
          
    CommentEntity:
      type: object
      properties:
        id:
          type: integer
          example: 1
        text:
          type: string
          example: This is a comment.
        authorId:
          type: integer
          example: 2
        filmId:
          type: integer
          example: 2
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T12:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-02T12:00:00Z
    
    CreateListDTO:
      type: object
      properties:
        name:
          type: string
          example: "My Favorite Movies"
        private:
          type: boolean
          example: true
      required:
        - name

    ChangeListVisibilityDTO:
      type: object
      properties:
        private:
          type: boolean
          example: true
      required:
        - private

    ListEntity:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "My Favorite Movies"
        private:
          type: boolean
          example: false
        authorId:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T12:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-02T12:00:00Z

    ShortInfoListWithAuthorAndFilms:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "My Favorite Movies"
        private:
          type: boolean
          example: false
        authorId:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T12:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-02T12:00:00Z
        author:
          type: object
          properties:
            name:
              type: string
              example: "username123"
        films:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 1
              posterUrlPreview:
                type: string
                example: https://placehold.co/600x400

    InfoListWithAuthorAndFilms:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "My Favorite Movies"
        private:
          type: boolean
          example: false
        authorId:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T12:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-02T12:00:00Z
        author:
          type: object
          properties:
            name:
              type: string
              example: "username123"
        films:
          type: array
          items:
            $ref: "#/components/schemas/FilmInListEntity"

    FilmInListEntity:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier of the film
          example: 1282979
        nameRu:
          type: string
          example: LEGO Ниндзяго
          nullable: true
        nameOriginal:
          type: string
          example: null
          nullable: true
        filmLength:
          type: integer
          description: Length of the film in minutes or in series
          example: 11
        year:
          type: integer
          example: 2019
        posterUrlPreview:
          type: string
          example: https://placehold.co/600x400
        createdAt:
          type: string
          format: date-time
          example: 2023-11-07T12:00:00Z
        type:
          type: string
          example: TV_SERIES
          enum:
          - VIDEO
          - FILM
          - MINI_SERIES
          - TV_SERIES
          - TV_SHOW
        countries:
          type: array
          example:
          - name: США
          - name: Канада
          - name: Дания
          items:
            type: object
            properties:
              name:
                type: string
        genres:
          type: array
          example:
          - name: приключения
          - name: боевик
          - name: мультфильм
          items:
            type: object
            properties:
              name:
                type: string
        avgRating:
          type: number
          format: float
          example: 6.0

  securitySchemes:
    AccessToken:
      type: apiKey
      name: Authorization
      in: header
    RefreshToken:
      type: apiKey
      name: Refresh Authorization
      in: header